Address PR review comments: decimal dtype, docstrings, dedup, notebooks
- Fix decimal() DSL helper to use DataType.DECIMAL instead of DOUBLE
- Fix weighted_sample_expr docstring to match actual behavior (missing
  weights get zero, not equal share)
- Remove duplicate _resolve_row_count from plan_builder, use shared
  parse_human_count from schema module
- Fix test_cdc_chunking.py docstring path
- Parameterize integration notebook hardcoded catalog/volume paths
  with dbutils.widgets
- Fix fk.py assert message field names (table_name, pk_column)
- Suppress ANN401 for _get_option (Any is intentional for v0 compat)

+++ dbldatagen/v1/dsl.py
@@ -85,7 +85,7 @@ def decimal(
 ) -> ColumnSpec:
     return ColumnSpec(
         name=name,
-        dtype=DataType.DOUBLE,
+        dtype=DataType.DECIMAL,
         gen=RangeColumn(min=min, max=max, **kw),
         seed_from=seed_from,
     )
